Revert to previous simpler previous SQL query
In commit https://github.com/MTG/freesound-datasets/commit/0a748ec3e8481cc1ca4625bced24e0aee9d059d0 we introduced a single SQL query that go num_ann, num_sounds and num_missing_votes in one go. However when tested in production we saw the query took hours to complete with full a sized dataset. To make it work in a reasonable amount of time we now do a query to get nun validated annotations for each node in the taxonomy. This should be refactored and use a single query to get all non validated annotation counts for all nodes.